Update indexer role to work with Ansible 12
Ansible 12 changes how conditionals work - https://docs.ansible.com/projects/ansible/latest/porting_guides/porting_guide_12.html

using the 4.14.1 branch and the ansible install guide in the documentation the indexer role fails when running the playbook.

> Error: [ERROR]: Task failed: Conditional result (False) was derived from value of type 'str' at '[...]/wazuh-ansible/roles/wazuh/wazuh-indexer/defaults/main.yml:35:22'. Conditionals must have a boolean result.

> Origin: [...]/wazuh-ansible/roles/wazuh/wazuh-indexer/tasks/security_actions.yml:117:7
>
>115   until: result.status in [200,201,401]
116   when:
117     - indexer_custom_user is defined and indexer_custom_user
          ^ column 7

suggested change is based on my conclusion, that the aim of the check is "exists and is not empty" (in contrast to "exists and is faulty"). I might be wrong.